rename apicultor to hub
(cherry picked from commit 82cc13cfb00cc47126eb72929c92fc4bb66ececc)
